feat: regular season games
While developing this, the NHL playoffs are ongoing. This means all the
data I use for recent games is playoff-focused.

This update now indicates what type of game is being played (i.e.,
regular season, playoffs, pre-season) as well as conditionally hides or
shows data based on the type.

For example, overtime losses matter in the regular season, but not in
the playoffs. It would be meaningless to show overtime losses when
showing a playoff game.